requirements.txt
requests==2.31.0
slack-sdk==3.27.1
notion-client==2.2.1
python-dotenv==1.0.1
